Make stop do nothing
In case monitoring fails the crm will try to stop/start -- it *will* try
to switch when starting, but not when stopping. So if the line fails we
won't have the off-on-off *and* we don't confuse the isdn-telecom switch
by a short burst of layer one from the other node.
When taking over the start command does the right thing (=fencing) and
will return an error if it doesn't see layer 1+2