ast_call
Optimized matching: A call is marked "False" (terminated) once we are
sure about our uniqueid. This avoids waiting for the last call that
started at the same time as our call.
Handler for NewAccountCode added and matching of account-code added: We
now match the account code (and set the "sure" flag) if the account is
specified. This is used by the RANDOMID feature of Call_Manager.