Add PIPE output to execute
Now we can add 'PIPE' output like for Subprocess.
Also add a communicate similar to Subprocess: In the process tree we
can have several 'PIPE' descriptors for stdout and stderr. These are
concatenated in tree order, see test4_exec.py for an example that
produces the first and the last line of a file by piping a 'cat' to
both, 'head' and 'tail'.
Note that communicate to do its work will add buffer processes to all
leaves of the tree (except for the first stderr) to avoid blocking.

To implement this, we now start subprocesses in depth-first order top
down. This might result in some processes getting a SIGCLD early if one
of the processes earlier in the start-sequence terminates early.